+++ src/test/java/com/inditex/pricing/unit/domain/valueobject/PriorityTest.java
@@ -9,7 +9,7 @@ class PriorityTest {
 
     @Test
     void createPrioritySuccessfully() {
-        long expectedValue = 1L;
+        int expectedValue = 1;
         Priority priority = Priority.of(expectedValue);
         assertNotNull(priority);
         assertEquals(expectedValue, priority.getValue());
@@ -17,15 +17,15 @@ class PriorityTest {
 
     @Test
     void checkNotEquals() {
-        Priority p1 = Priority.of(1L);
-        Priority p2 = Priority.of(2L);
+        Priority p1 = Priority.of(1);
+        Priority p2 = Priority.of(2);
         assertNotEquals(p1.getValue(), p2.getValue());
     }
 
     @Test
     void checkEquals() {
-        Priority p1 = Priority.of(5L);
-        Priority p2 = Priority.of(5L);
+        Priority p1 = Priority.of(5);
+        Priority p2 = Priority.of(5);
         assertEquals(p1.getValue(), p2.getValue());
         assertNotSame(p1, p2);
     }
